# Faster nibble parsing and a SSSE3 routine for even faster nibble parsing