Nibble parsing was written as a for loop. This leads to an extra loop control variable being used. By using pointer arithmetic, the loop control variable can be parsed out. This leads to faster code.

I also wrote a routine to utilise the pshufb instruction in order to get even faster translation.

Speed before:
```
Benchmark 1: ./test/bin/htsfile ~/test/nanopore_100000reads.u.bam  --view > /dev/null
  Time (mean ± σ):     893.6 ms ±   9.7 ms    [User: 770.5 ms, System: 122.9 ms]
  Range (min … max):   878.6 ms … 913.4 ms    10 runs
```

Speed after:
```
Benchmark 1: ./test/bin/htsfile ~/test/nanopore_100000reads.u.bam  --view > /dev/null
  Time (mean ± σ):     694.4 ms ±   6.7 ms    [User: 583.3 ms, System: 110.6 ms]
  Range (min … max):   682.1 ms … 703.0 ms    10 runs
```

Speed after, using SSSE3
```
Benchmark 1: ./test/bin/htsfile ~/test/nanopore_100000reads.u.bam  --view > /dev/null
  Time (mean ± σ):     524.7 ms ±  11.9 ms    [User: 400.3 ms, System: 124.3 ms]
  Range (min … max):   510.0 ms … 552.7 ms    10 runs
```

Both the non-ssse3 and ssse3 function deliver the same htsfile output for all the 100.000 reads tested. (Tested using a checksum.)