@daviesrob 2023-09-26
This looks interesting, but it would need some form of dynamic dispatch.  While SSSE3 may be widespread now, we can't make any assumptions about the hardware in use.