@rhpvorderman 2023-09-26
@daviesrob Sure. I have been meaning to practice with dynamic dispatch for sometime. At least it is also quite a lot faster without the SSSE3 routine. Would it be an idea to remove the SSSE3 code from this PR and do that later in a separate PR?