@rhpvorderman 2023-09-26
Strangely enough I retested on a intel core i7 -6700 system and the result was slower than on my ryzen 5 5650 system. So it seems to differ per CPU strangely enough whether it is faster or slower using the while loop vs the for loop.

Here is the loop in assembly for this PR
```Assembly
.L3:
        movzx   esi, BYTE PTR [rcx]
        add     rax, 2
        add     rcx, 1
        movzx   esi, WORD PTR code2base.0[rsi+rsi]
        mov     WORD PTR [rax-2], si
        cmp     rax, r8
```
Develop branch
```
.L3:
        movzx   esi, BYTE PTR [rdi+rax]
        movzx   esi, WORD PTR code2base.0[rsi+rsi]
        mov     WORD PTR [rbp+0+rax*2], si
        add     rax, 1
        cmp     ecx, eax
        jg      .L3
 ```
 I can't see why one branch would be faster than the other.

The SSSE3 routine is always faster, but indeed that requires some extra work.