@jkbonfield 2023-09-26
I was testing on Intel and found it's also highly dependent on compiler, and probably version.

Original scalar code, for 100k ONT reads (NB this nibble function only, perf counts):  clan16 543, gcc13 413
Your pointer based scalar code: clang16 314, gcc13 553.
An unrolled alternative: clang16 325, gcc13 368.

The SSSE3 implementation of this function however was around 110 perf hits/counts for both compilers.
However I'm unsure if the code is worth it as for me at least it's only looking like ~4% speed gain to the overall test_view program.  What was in your BAM file?  Is it just sequence, or also having quality values, alignments, aux tags, etc?  Also what compiler and version are you using?  What CPU?

The unrolled alternative looks better than both orig and the new variant as far as scalar code goes.  It's a work-in-progress, but I have atm:

```
    const char *seq_end_ptr_8atatime = seq + (len & (~7ULL));                                        
    uint16_t *c16 = (uint16_t *)code2base;                                                           
    while (seq_cursor < seq_end_ptr_8atatime) {                                                      
        memcpy(seq_cursor+0*2, &c16[nibble_cursor[0]], 2);                                           
        memcpy(seq_cursor+1*2, &c16[nibble_cursor[1]], 2);                                           
        memcpy(seq_cursor+2*2, &c16[nibble_cursor[2]], 2);                                           
        memcpy(seq_cursor+3*2, &c16[nibble_cursor[3]], 2);                                           
        memcpy(seq_cursor+4*2, &c16[nibble_cursor[4]], 2);                                           
        memcpy(seq_cursor+5*2, &c16[nibble_cursor[5]], 2);                                           
        memcpy(seq_cursor+6*2, &c16[nibble_cursor[6]], 2);                                           
        memcpy(seq_cursor+7*2, &c16[nibble_cursor[7]], 2);                                           
                                                                                        
        seq_cursor += 16;                                                                            
        nibble_cursor += 8;                                                                          
    }                                                                                                
```

I can tidy that up a bit more as it has remnant of earlier experiments in there too with uint16_ts.

I can test AMD too though as it sounds like that's different again.