@rhpvorderman 2023-09-27
The reads in this set are very long. They are from the GM24385_1.fastq.gz file (from the GIAB data indices, extra long ONT reads). The average read length was longer than 7000. This data was converted to BAM using Picard FastqToSam, in order to simulate uBAM. Because of the very long reads I think the algorithm has a chance to truly shine.

I wrote a little uBAM-only parser for dnaio (alignments are ignored) and found that the SSSE3 routine improved parsing speed about 20%, so that is concordant with your result. If I cheat and do not store the BAM in bgzip blocks (just decompress the bam with gzip), then the speedup is 33%. In my uBAM parser I also used SSE2 for adding +33 to all the qualities so that is less of a bottleneck. 

How should I proceed from this point? It is clear that the scalar code in this PR is mixed depending on hardware (so not worth it) and the SSSE3 code is phenomally better, but also much harder to integrate. 10-15% overall application performance by improving a single function is quite nice. So it would feel a bit wasteful not to do anything with it.