@jkbonfield 2023-09-27
My ONT file had an average read length of about 4.6kb, so similar ballpark.  You probably gained more because you didn't have any alignment fields or aux tags to print up, meaning sequence printing becomes more significant.

Agreed a 10-15% speed gain for this case is worth having, even more so if the circumstances make that a larger percentage.  I'm not sure it's worth it if it's only included if the user builds using `-march=native` or `-mssse3` though.

I haven't tried it, but there is this: https://gcc.gnu.org/wiki/FunctionMultiVersioning and more blurb at https://johnnysswlab.com/cpu-dispatching-make-your-code-both-portable-and-fast/

It'd need appropriate compiler guards to ensure it's only there on compilers that support it (see htslib/hts_defs.h for examples).

For htscodecs [I went with the more complex dispatching route](https://github.com/samtools/htscodecs/blob/master/htscodecs/rANS_static4x16pr.c#L833-L1124), but primarily because when I started it I wasn't aware of the builtin methods.   Since then though it's turned out to be helpful because I have multiple versions and it's not always best to use the highest SIMD implementation on all platforms.  That's over the top though for this function I think.

Regarding the qual+33, I'm surprised it's not auto-vectorised.  If not then maybe this is a case where we ought to be using `restrict`.